promptdb/requirements.txt
mysql-connector-python~=9.0.0
requests~=2.32.3
tabulate~=0.9.0
setuptools~=68.1.2